Used deferred execution in Sequence methods
Methods like Sequence.where and Sequence.map are here implemented using
deferred execution. They return an instance of a new Sequence-derived class
that performs the operation while iterating. I've also added Sequence.skip and
Sequence.take in similar fashion.

In order to still provide a way to easily get a new list with the result,
Sequence.toList was added.

The advantage of this approach is that it avoids the allocation of intermediate
lists. The following example allocates 4 small Sequence-derived classes, but
only a single list:

  IO.print((1..100000000).
           where { |i| i % 2 == 0 }.
           map { |i| i*i }.
           skip(1).take(3).toList())

  // expect: [16, 36, 64]