Add syntax for named function declarations.
This is technically redundant since it's equivalent to:

var f = Fn.new { ... }

But it seems worth adding if we're going to make functions and a
functional/procedural style intuitive and natural in Wren. It's a bit
of sugar, but it doesn't hurt much to add and makes the code read a
lot better.