Allow "var" in a class body for defining "properties".
A property is a field with an implicit getter, setter, and optional
class body initializer.

It's handy for defining publicly visible state in a class. When modules
are classes, this is needed for "top level" variables.

Right now, a class var gets both a getter and setter. It would be nice
to also have something like "val" for properties that are publicly
visible but not settable.

Also, still need to support "static var" for metaclass properties.