Make wrenCall operate on a stack.
This breaks previous way of working. Before this change when creating a
wrenCall, the stack was ignoring top unwanted elements when invoking a
function. So if you had a stack like:

  foo | bar | baz

calling 'foo(_)' would tune the stack as:

  foo | bar

and make 'foo' the receiver and 'bar' the argument for 'foo(_)'.

After this change, it will work as a regular stack. So it means that the
stack will stay unchanged, and 'bar' will be the receiver and 'bar' the
argument for 'foo(_)'.

While there are pro's for the simplicity of previous way to access the
arguments, the new way should empower more the users. By giving finer grained
control of the stack when making consecutive wrenCall's. It is now possible to
reserve some slots at the begin of the stack, to store Fiber contextual
variables to be reused on wrenCall.