snap: Add snapcraft.yaml
This commit introduces support the bundling/delivering wren through
the 'snap' mechanism/store - for more details see https://snapcraft.io.

"A snap is a bundle of your app and its dependencies that works without
modification across Linux.  Snaps are discoverable and installable from
the Snap store, an app store with an audience of millions." [1]

In case the support/snapcraft.yaml file is accepted it can be published
in the Snap Store [2] which is required for people to easily install it.

Assistance for publishing in the snap store is available if required.

Test Procedure
==============

Build/Snap
----------

$ lxc launch ubuntu:16.04 snapcrafting  # create 'snapcrafting' container
$ lxc exec snapcrafting -- su - ubuntu  # execute user shell in container

$ sudo snap install snapcraft --classic # install snapcraft tool/commands
$ sudo apt-get update # update for installing build dependencies packages

$ git clone https://github.com/wren-lang/wren
$ cd wren

$ # get this snapcraft.yaml file
$ snapcraft
...
Building wren

make -j4

...
Snapped wren_0.1.0_amd64.snap

Install/Contents
----------------

$ sudo snap install wren_0.1.0_amd64.snap --dangerous
$ # '--dangerous' is required as this isn't from the store.

$ which wren
/snap/bin/wren

Testing
-------
Create a file named “my_script.wren” in your favorite text editor and paste this into it:
...
for (yPixel in 0...24) {
  var y = yPixel / 12 - 1
  for (xPixel in 0...80) {
    var x = xPixel / 30 - 2
    var x0 = x
    var y0 = y
    var iter = 0
    while (iter < 11 && x0 * x0 + y0 * y0 <= 4) {
      var x1 = (x0 * x0) - (y0 * y0) + x
      var y1 = 2 * x0 * y0 + y
      x0 = x1
      y0 = y1
      iter = iter + 1
    }
    System.write(" .-:;+=xX$& "[iter])
  }

  System.print("")
}
...

Then run:
$ wren my_script.wren

References:
[1] "Creating a snap" in https://docs.snapcraft.io/creating-a-snap/6799
[2] "Share with your friends" in https://docs.snapcraft.io/c-c-applications/7817

Signed-off-by: Dongdong Tao  <dongdong.tao@canonical.com>