wip: initial premake script, which generates the wren static library + wren_test binary
This will isolate the VM from the CLI by making the tests self contained. Note that the intention is to commit ready to build project files for vs, xcode, and make such that there is no dependency or steps in between to build a usable wren library.  Also note the goal is not less portability, just a more reliable process for the 99% of users looking to try out the language.