Allow trailing comma on methods
That includes declaration and call, both methods, subscripts and block arguments.

The reasons are: a) we allow trailing commas on lists and maps, and b) git diff on multiline methods.

As a side benefit, we also get a more user-friendly error message in case of empty subscript or block argument's parameters list (`Fn.new {||}`).