Report a runtime error when found an undefined foreign class
In the previous code, such class made `ASSERT()` fail in debug mode, and triggered a segfault in production, in the better case. In the worse case, it was a security hole (UB) and all sorts of bad.
It was one of two - dereferencing an uninitialized pointer or out-of-bounds array access. Neither way a good thing, and security men are expert in seriously crafting code so it will hit the out-of-bounds case and to some sensitive data. We're all hate it.