Tentative fix for Num class docs issue on the wren.io site.
I noticed today that, in the documentation for the ```Num``` class on the wren.io site, the explanation for the ```<<``` operator is not being displayed (at least on Firefox and Safari). It is being overwritten by the explanation for the ```..``` operator which, however, has lost its heading.

Although at first sight the markdown file looks OK, I suspect the reason for this is that we should be using HTML escapes for the ```>```characters  (and also for that matter the ```<``` characters). If so, this is my bad as I prepared the original PR when these docs were added.

I've therefore changed them in the hope this will fix the problem.