Add `Object.hash`
The basic rationnale is to offer a fast access to the hash of the
interal objects, so users can implement their containers more easily and
efficiently.

In addition, it also allow regular objects to be hashed by their reference. It
allows to have access to a fast hash for objects meant to be unique
instances (where `==` does not need to be overloaded).

`Map` behaviour is preserved by `wrenMapIsValidKey` checking we don't
try to pass invalid key values. (Not preserving this behaviour would be
to complex, and would require reentrancy...)