snapshot restore: avoid wrenDumpCode() when any swizzle is found
wrenDumpCode() does ASSERT() on the UNDEFINED_VAL which means a to-be-swizzled
thing.

Let's recognize any swizzle as a reason to avoid calling this function.

In practice, in the buffer of constants of an ObjFn being restored, only the
ObjFn type was found to be swizzled.
In the core module, the List.sort() is an example.